+++ aicoe/sesheta/actions/__init__.py
@@ -1,0 +1,4 @@
+"""Sesheta's actions."""
+
+__title__ = "sesheta-actions"
+__version__ = "0.1.0"